+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## Version 1.0.1
+
+  * Fix Osm::Term date helping methods when term contains nil dates
+  * Fix fetching event attendance when no attendance has been set for any members
+  * Osm::Api exposes the debug option as a psudo class attribute
+
 ## Version 1.0.0
 
   * SMS text messages: